Subjects now disconnects observers in the reversed order of connection.
Observers that are connected as last often have a shorter lifetime than
those which are connected first. Calling disconnect of its observers in
the reversed order in which they are connected reduces issues when there
are dependencies between observers.

The added test ensures the order of notification and disconnecting the
observers.